+++ bm-wp-experience.php
@@ -173,6 +173,7 @@ class BM_WP_Experience {
 		require_once 'includes/class-plugins.php';
 		require_once 'includes/class-rest-api.php';
 		require_once 'includes/class-security.php';
+		require_once 'includes/class-updates.php';
 		require_once 'includes/class-users.php';
 	}