+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file. This projec
 - Remove color scheme picker by default. Can be overridden via filter `bm_wpexp_remove_color_scheme_picker`.
 - Disable comments by default. Can be overridden by defining `BM_WP_ENABLE_COMMENTS` as true in the config.
 - Remove Yoast SEO dashboard widget.
+- Feature to disable update notices when website is on maintenance plan. To use, set `BM_WP_HAS_MAINTENANCE_PLAN` to true.
 
 ### Changed